ci(accumulator): move `pop_unsaved` test to moveos-commons
Relocated the `accumulator_pop_unsaved` test from `rooch-store` to `moveos-commons`. This improves alignment with test modularity by grouping it with accumulator-specific logic within `moveos-commons`.
